New dev

+++ forum/api/views.py
@@ -2,17 +2,24 @@ from django.shortcuts import render
 from django.http import HttpResponse
 import json
 import MySQLdb
+from data_service.user import *
+
 from django.utils.datastructures import MultiValueDictKeyError
 
 db = MySQLdb.connect(host="localhost", user="AlexO", passwd="pwd", db="tp_project_forum")
 
 def user_create(request):
     if request.method == "POST":
-        username = request.POST['username']
-        about = request.POST['about']
-        isAnonymous = request.POST['isAnonymous']
-        name = request.POST['name']
-        email = request.POST['email']
+        try:
+            username = request.POST['username']
+            about = request.POST['about']
+            isAnonymous = request.POST['isAnonymous']
+            name = request.POST['name']
+            email = request.POST['email']
+        except MultiValueDictKeyError:
+            result = {}
+            return HttpResponse(json.dumps(result), content_type='application/json')
+
         cursor = db.cursor()
         try:
             cursor.execute("""INSERT INTO Users (username, email, name, isAnonymous, about)
@@ -34,73 +41,45 @@ def user_create(request):
         return HttpResponse(json.dumps(result), content_type='application/json')
     return HttpResponse(json.dumps(0), content_type='application/json')
 
-
-def user_info(email):
-    cursor = db.cursor()
-    cursor.execute("""SELECT * FROM Users WHERE email = %s; """, (email,))
-    resExec = cursor.fetchall()
-
-    cursor.execute("""SELECT t2.email FROM Users AS t1
-                          INNER JOIN Users_has_Users AS t ON t.Users_id = t1.id
-                          INNER JOIN Users AS t2 ON t.Users_id1 = t2.id
-                          WHERE t1.email = %s AND t1.id != t2.id""", (email,))
-    followers = cursor.fetchall()
-
-    count = cursor.execute("""SELECT t2.email FROM Users AS t1
-                                  INNER JOIN Users_has_Users AS t ON t.Users_id1 = t1.id
-                                  INNER JOIN Users AS t2 ON t.Users_id = t2.id
-                                  WHERE t1.email = %s AND t1.id != t2.id""", (email,))
-    following = cursor.fetchall()
-
-    result = {'about': resExec[0][5],
-              'email': resExec[0][2],
-              'followers': followers,
-              'following': following,
-              'id': resExec[0][0],
-              'isAnonymous': resExec[0][4],
-              'name': resExec[0][3],
-              'subscriptions': count,
-              'username': resExec[0][1]}
-    cursor.close()
-    return result
-
-def informationAboutUser(user_id, flag):
-    email = getUserEmailByID(user_id)
-    if flag:
-        return user_info(email)
-    else:
-        return email
-
-
 def user_details(request):
     if request.method == "GET":
-        email = request.GET['user']
+        try:
+            email = request.GET['user']
+        except MultiValueDictKeyError:
+            result = {}
+            return HttpResponse(json.dumps(result), content_type='application/json')
+
         result = user_info(email)
         return HttpResponse(json.dumps(result), content_type='application/json')
     return HttpResponse(json.dumps(0), content_type='application/json')
 
 def user_follow(request):
-    follower = request.POST['follower']
-    followee = request.POST['followee']
-    #follower = "user1@yandex.ru"
-    #followee = "alex@yandex.ru"
+    #follower = request.POST['follower']
+    #followee = request.POST['followee']
+
+    follower = "user1@mail.ru"
+    followee = "user2@mail.ru"
+
 
     cursor = db.cursor()
 
-    cursor.execute("""SELECT id FROM Users WHERE email = %s; """, (follower,))
-    follower_id = cursor.fetchall()
+    follower_id = getUserIDByEmail(follower)
+    if follower_id is None:
+        return HttpResponse(json.dumps(0), content_type='application/json')
 
-    cursor.execute("""SELECT id FROM Users WHERE email = %s; """, (followee,))
-    followee_id = cursor.fetchall()
+    followee_id = getUserIDByEmail(followee)
+    if followee_id is None:
+        return HttpResponse(json.dumps(0), content_type='application/json')
 
     try:
         cursor.execute("""INSERT INTO Users_has_Users (Users_id, Users_id1)
                              VALUES (%s, %s);""", (followee_id, follower_id))
+        db.commit()
+        result = user_info(follower)
     except MySQLdb.Error, e:
-        print "User exist"
+        db.rollback()
+        result = {}
 
-    db.commit()
-    result = user_info(follower)
     return HttpResponse(json.dumps(result), content_type='application/json')
 
 def user_listFollowers(request):
@@ -145,31 +124,32 @@ def user_listFollowing(request):
     return HttpResponse(json.dumps(result), content_type='application/json')
 
 def user_unfollow(request):
-    follower = request.POST['follower']
-    followee = request.POST['followee']
-    #followee = "email"
-    #follower = "alex@yandex.ru"
+    #follower = request.POST['follower']
+    #followee = request.POST['followee']
+
+    follower = "user1@mail.ru"
+    followee = "user2@mail.ru"
 
     cursor = db.cursor()
 
-    cursor.execute("""SELECT id FROM Users WHERE email = %s; """, (follower,))
-    follower_id = cursor.fetchall()
+    follower_id = getUserIDByEmail(follower)
+    if follower_id is None:
+        return HttpResponse(json.dumps(0), content_type='application/json')
 
-    cursor.execute("""SELECT id FROM Users WHERE email = %s; """, (followee,))
-    followee_id = cursor.fetchall()
+    followee_id = getUserIDByEmail(followee)
+    if followee_id is None:
+        return HttpResponse(json.dumps(0), content_type='application/json')
 
     try:
         cursor.execute("""delete from Users_has_Users where
                           users_id = %s and users_id1 = %s""", (followee_id, follower_id))
         db.commit()
-    except MySQLdb.Error, e:
+        result = user_info(follower)
+        cursor.close()
+        return HttpResponse(json.dumps(result), content_type='application/json')
+    except MySQLdb.Error:
         db.rollback()
-        print "No follower"
-
-    result = user_info(follower)
-
-    cursor.close()
-    return HttpResponse(json.dumps(result), content_type='application/json')
+        return HttpResponse(json.dumps(0), content_type='application/json')
 
 def user_updateProfile(request):
     about = request.POST['about']
